src/scc-slack-cloud-function/requirements.txt
requests
google-cloud-secret-manager
google-cloud-logging